feat: 댓글 전체 조회

+++ src/main/java/likelion14th/springstudy/service/ArticleService.java
@@ -36,14 +36,22 @@ public class ArticleService {
         return articleResponses;
     }
     @Transactional
-    public ArticleDetailResponse patchArticle(Long Id,String title, String content){
+    public ArticleDetailResponse patchArticle(Long Id,String title, String content, String password){
         Article article=articleRepository.findById(Id).orElseThrow(()->new EntityNotFoundException("해당 Id의 게시글이 없습니다"));
+        if (!article.getPassword().equals(password)){
+            throw new RuntimeException("해당 게시글에 대한 수정 권한이 없습니다.");
+        }
         article.update(title,content);
         articleRepository.save(article);
         return ArticleDetailResponse.from(article);
     }
     @Transactional
-    public Void deleteArticle(Long Id){
+    public Void deleteArticle(Long Id, String password){
+        Article article = articleRepository.findById(Id)
+                .orElseThrow(() -> new EntityNotFoundException("해당 ID의 게시글을 찾을 수 없습니다."));
+        if (!article.getPassword().equals(password)){
+            throw new RuntimeException("해당 게시글에 대한 삭제 권한이 없습니다.");
+        }
         articleRepository.deleteById(Id);
         return null;
     }
